requirements.txt
asgiref==3.3.1
cffi==1.14.4
cryptography==2.9.2
Django==3.1.5
django-environ==0.4.5
entrypoints==0.3
jeepney==0.6.0
keyring==18.0.1
lxml==4.5.2
pycparser==2.20
Pygments==2.3.1
pytz==2020.5
SecretStorage==3.3.0
six==1.15.0
sqlparse==0.4.1
pip~=20.0.2
wheel~=0.34.2
beautifulsoup4~=4.8.2
html5lib~=1.0.1
psycopg2~=2.8.5
PyYAML~=5.3.1
Jinja2~=2.11.2
geoip2~=4.0.2
numpy~=1.17.4
simplejson~=3.16.0
Pillow~=7.0.0
setuptools~=45.2.0
requests~=2.7.0
Werkzeug~=0.16.1
djangorestframework~=3.11.0